platform/x86/intel/pmt: Add Watcher driver
INTEL INTERNAL ONLY

Watcher is a capability of the Intel Platform Monitoring Technology.
The Watcher capability provides access to hardware sampling and tracing
functionality.

PMT architecture uses PCI to enumerate the Watcher capability and
to access configuration registers and result buffers. Watcher supports
two types of hardware collection features, Sampler and Tracer. Both
Sampler and Tracer allow configuring a collection mode (oneshot,
periodic), a period, and an enabling vector to select the metrics to
be sampled/traced. On Intel devices, Tracer result data is processed
by the CPU Trace Hub hardware, while Sampler data is made available to
the OS through MMIO.

The PMT Watcher capability has a connection with the PMT Telemetry
capability (see Intel PMT Watcher capability driver). It is from a
Telemetry region that Watchers are configured to be sampled or traced
from. Like Telemetry, Watcher also provides a GUID that is used by
software to identify a version of a XML. Though related to the Telemetry
capability, Watchers function as stand alone components that do not
interact with or otherwise disturb the operation of a Telemetry.

Tracers and Samplers are enumerated separately in the class folder.
Files to configure a Tracer or Sampler are provided along with a
pointer to the parent device the feature belongs to. For samplers,
a binary sysfs attribute is created that allows user space to mmap
the result buffer.

Also move early_client_hw pci id list to the intel_pmt_class.c since
the issues are common to the discovery of both Telemetry and Watcher
features.

Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: "David E. Box" <david.e.box@linux.intel.com>