x86/cpu: Enable HRESET
Enabling the HRESET instructions involves several steps. Once support for
the instruction has been identified, inspect which HRESET features are
supported in the associated CPUID leaf. Then, program the
IA32_HW_RESET_ENABLE register with the desired features. The history
features to be clobbered when calling HRESET are specified in %eax.

Enable all the supported HRESET features if the instruction is supported.
Also, clobber all history predictions when calling the instruction. For
this purpose save the supported features a global variable that can be
used in entry code.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>