KVM: VMX: Support virtual HFI table for VM
The Hardware Feedback Interface (HFI) is a feature that allows hardware
to provide guidance to the operating system scheduler through a hardware
feedback interface structure (HFI table) in memory [1], so that the
scheduler can perform optimal workload scheduling.

ITD (Intel Thread Director) and HFI features both depend on the HFI
table, but their HFI tables are slightly different. The HFI table
provided by the ITD feature has 4 classes (in terms of more columns
in the table) and the native HFI feature supports 1 class [2].

In fact, the size of the HFI table is determined by the feature bit that
the processor supports, but the range of data updates in the table
is determined by the feature actually enabled (HFI or ITD) [3], which is
controlled by MSR_IA32_HW_FEEDBACK_CONFIG.

To benefit the scheduling in VM with HFI/ITD, we need to maintain
virtual HFI tables in KVM. The virtual HFI table is based on the real
HFI table. We extract the HFI entries corresponding to the pCPU that the
vCPU is running on, and reorganize these actual entries into a new
virtual HFI table with the vCPU's HFI index.

Also, to simplify the logic, before the emulation of ITD is supported,
we build virtual HFI table based on HFI feature by default (i.e. only 1
class is supported, based on class 0 of real hardware).

Add the following support:

- Thermal interrupt injection.
  The update of the HFI table is notified via the thermal interrupt.
  Therefore, once the virtual HFI table is updated, we need to inject
  thermal interrupts into the corresponding VMs.

- Interfaces to initialize and build the virtual HFI table for the VM.
  Once the HFI/ITD is enabled or the table address changes, we need to
  initialize/build the entire HFI table. These VM level interfaces are
  prepared for the following HFI/ITD related MSRs emulation.

[1]: SDM, vol. 3B, section 15.6 HARDWARE FEEDBACK INTERFACE AND INTEL
     THREAD DIRECTOR
[2]: SDM, vol. 3B, section 15.6.2 Intel Thread Director Table Structure
[3]: SDM, vol. 3B, section 15.6.5 Hardware Feedback Interface
     Configuration, Table 15-10. IA32_HW_FEEDBACK_CONFIG Control Option

Signed-off-by: Zhuocheng Ding <zhuocheng.ding@intel.com>
Co-developed-by: Zhao Liu <zhao1.liu@intel.com>
Signed-off-by: Zhao Liu <zhao1.liu@intel.com>
---
Changes since v3:
 * None.

Changes since v2:
 * Split per-vCPU hfi_table_idx and per-VM hfi_desc supports into
   separate patches to simplify the code logic.
 * Rename the title from "KVM: x86: *" to "KVM: VMX: *".

Changes since v1:
 * Rename vmx_hfi_is_initialized() and vmx_hfi_int_is_enabled() to
   vmx_hfi_initialized() and vmx_hfi_int_enabled() to match the naming
   style.