KVM: VMX: Emulate the MSR of HRESET feature
HRESET is a feature associated with ITD, which provides an HRESET
instruction to reset the ITD related history accumulated on the current
logical processor it is executing on [1]. The HRESET instruction does
not cause the VMExit and is therefore available to the Guest by default
when the HRESET feature bit is set for the Guest.

The HRESET feature also provides a thread scope MSR to control the
enabling of the ITD history reset via the HRESET instruction [2]:
MSR_IA32_HW_HRESET_ENABLE.

This MSR can control the hardware, so we use the emulation way to
support it for Guest, and this makes the Guest's changes to the hardware
under the control of the Host.

Considering that there may be the difference between Guest and Host
about HRESET enabling status, we store the MSR_IA32_HW_HRESET_ENABLE
values of Host and Guest in vcpu_vmx and save/load their respective
configurations when Guest/Host switch.

[1]: SDM, vol. 3B, section 15.6.11 Logical Processor Scope History
[2]: SDM, vol. 2A, chap. CPUID--CPU Identification, CPUID.07H.01H.EAX
     [Bit 22], HRESET.

Co-developed-by: Zhuocheng Ding <zhuocheng.ding@intel.com>
Signed-off-by: Zhuocheng Ding <zhuocheng.ding@intel.com>
Signed-off-by: Zhao Liu <zhao1.liu@intel.com>
---
Changes since v3:
 * Drop the "itd_guest_{enter|exit}" and "hfi_guest_{enter|exit}"
   changes.
 * Add the reserved bits check for MSR_IA32_HW_HRESET_ENABLE emulation.

Changes since v2:
 * Don't save/load MSR_IA32_HW_HRESET_ENABLE when Host and Guest have
   the same MSR value.
 * Emulate MSR_IA32_HW_HRESET_ENABLE other than passing-through it to
   Guest.
 * Rename the title from "KVM: x86: *" to "KVM: VMX: *".
 * Split the HRESET CPUID support into the separate patch.
 * Delete the comment in hreset_guest_enter() since the reason why we
   don't reset history at VM entry is the most timer interrupts are not
   posted interrupt and their injection will cause VM exit so that
   reseting will clear the history before we do classification.

Changes since v1:
 * Remove the emulation check of MSR_IA32_HW_HRESET_ENABLE in
   vmx_has_emulated_msr() since this MSR is passed-through.
 * Remove MSR_IA32_HW_HRESET_ENABLE from msrs_to_save_base[] since it
   is passed-through to Guest and there's no need to expose it to VMM.
 * Rewrite the comment in hreset_guest_enter() to make it clearer.