mm/pkeys: Provide for PKS key allocation
Kernel consumers of PKS need a way to allocate a PKS pkey and assign the
initial permissions for that key.  It is desirable to not allocate keys
for consumers which are not configured.

Introduce a macro to allocate keys sequentially based on which consumers
are configured.  In addition define a macro to set the proper permission
bits based on the actual pkey value allocated.

pks-keys.h is added as a new header with minimal header dependencies.
This allows the use of PKS_INIT_VALUE within other headers where the
additional includes from other pkey headers caused major conflicts.  The
main conflict was using PKS_INIT_VALUE for INIT_TRHEAD in
asm/processor.h

Add documentation.

Suggested-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>