[REVERTME] KVM: TDX: Fixup fatal #MC happens in TD guest that delivered as SMI
When BIOS eMCA MCE-SMI morphing is enabled, the #MC happens in TD
guest is morphed to SMI. Then the SMI causes TD exit reason (6) to KVM.

Once SEAMCALL(TDENTER) returns from SEAM module, the SMI is delivered
and bios SMI handler is called. After some handling, the SMI handler
delivers #MC to OS and returns back to OS. In this case, host kernel #MC
handler has no idea of whether the #MC is caused by a TD guest. So if
the #MC is fatal type, it will panic the kernel. We really don't want
this.

Note, it's different from the case that TD exit due to #MC. In that case,
KVM calls kvm_machine_check(), which fakes a ring 3 pt_regs so that
fatal machine just causes user space VMM to be killed.

To survive from the fatal #MC deliverd as MSMI->MCE, that happens in TD
guest. Put a fixup entry after SEAMCALL(TDENTER) and also expand the
condition to set MCE_IN_KERNEL_RECOV to make this hack work.

TODO: make this patch upstreamable

Signed-off-by: Xiaoyao Li <xiaoyao.li@intel.com>
Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>