x86/cpu: Add a generic helper function to get the full hybrid CPU parameters
On a hybrid processor, each CPU has a u32 parameter that is composed by
both the CPU type (in bits [31:24]) and the native model ID (in bits
[23:0]). On some processors, there may be CPUs of the same type but with
different native mode IDs. Thus, to fully identify a CPU, it is necessary
to query both parameters.

In preparation for a subsequent changeset that returns both combined
parameters. Introduce a new helper function __do_get_hybrid_params() that
reads the mentioned parameters from CPUID. Implement the function in such
a manner that can be called directly and on a different CPU via an IPI.

Rework get_this_hybrid_cpu_type() to make use of this new function.

Cc: Andi Kleen <ak@linux.intel.com>
Cc: Andy Lutomirski <luto@kernel.org>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Kan Liang <kan.liang@linux.intel.com>
Cc: "Peter Zijlstra (Intel)" <peterz@infradead.org>
Cc: "Rafael J. Wysocki" <rafael.j.wysocki@intel.com>
Cc: "Ravi V. Shankar" <ravi.v.shankar@intel.com>
Cc: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Cc: linux-kernel@vger.kernel.org
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>