x86/cpu: Restore MSR_IA32_HW_HRESET_ENABLE MSR on BSP during wakeup
The value of MSR_IA32_HW_HRESET_ENABLE is lost when suspending. Given that
HRESET is executed when switching to kernel mode and viceversa, a general
protection fault occurs when HRESET is executed for the first time after
waking up. Thus, restore such MSR when waking up. There is no need to
restore this MSR for APs because it will be restored during CPU hotplug
online.

[Ricardo edited the commit message for clarity.]

Reported-by: Wendy Wang <wendy.wang@intel.com>
Reported-by: kernel test robot <lkp@intel.com>
Co-developed-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>