x86/tdx: Block Xen initialization in TDX
Xen has a lot of complicated code that would be a lot of work to audit
and harden. Right now nobody is planning to enable TDX on Xen, so it
seems safer to just disable it. In principle a malicious hypervisor
could fake being Xen and then use a potential hole in the Xen drivers
to attack the guest.

If Xen ever gets enabled for TDX this can be removed, but of course in
this case the drivers and other Xen code would need to be audited and
hardened first.

Add a CPUID cc_platform_has filter and use that to disallow Xen HVM
initialization. For Xen PVH it's impossible anyways because that
requires a special entry point that even a malicious hypervisor
couldn't call.

Cc: jgross@suse.com
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>