x86/tdx: Enable panic_on_oops by default in TDX
For a confidential guest it seems safer to always panic on oopses.
Some kind of attack from the host might result in a oops, and
the oops recovery path can leave kernel subsystems in various
undefined states, which could be used by the attacker.

This can be still overridden after boot using the sysctl.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>