ioasid: Fix unexpected ioasid_set free
This patch fixes a problem in which ioasid_set is free when the last ioasid
in it is free. I think the motivation of this atomatic ioasid_set free is
based on the consideration when an ioasid_set_free() is called but there is
still ioasid within it. To reclaim the ioasid_set later, this logic was added.

Cc: Jacob Pan <jacob.jun.pan@intel.com>
Signed-off-by: Liu Yi L <yi.l.liu@intel.com>