platform/x86: Add RTMR update interface
Add an experimental interface to update the RTMR measurement registers
in the TDX module from ring 3. Simply allow to pass in a update hash
and pass it on to a TDCALL. The interface has the same rights
as any other operation on the tdx-attest device node, so becomes
controllable by udev rules.

This is currently not for upstream, but mainly for internal
testing.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>