random: Force trusting CPU RNG in TDX
In TDX relying on environmental noise like interrupt timing alone is
dubious, because it can be directly controlled by a untrusted
hypervisor. Make sure to always mix in the CPU hardware random number
generator too.

Do this by forcing to trust the CPU random's generator when running
inside TDX. In this case the random number generator needs to be trusted
anyways because all the keys for memory encryption and attestation depend
on it. This way the default CONFIG setting can be still used
in the same kernel binary when TDX is not used.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>