acpi: Make acpi_scan_handler::detach to return error
Because TDX(according to TDX module specification 344425-002US) doesn't
support CPU hotplug, physical CPU hot unplug must be disabled when TDX is
enabled.  The acpi_scan_handler.detach method that is called to unplug
devices doesn't return a value.  Make the method return a value to allow
refusing a physical unplug, and make the detach method for the processor
allow this refusal.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>